Move model to src root
