+++ lib/src/main/java/com/zetcode/model/DocumentModelEx.java
@@ -1,4 +1,4 @@
-package com.zetcode;
+package com.zetcode.model;
 
 import javax.swing.BorderFactory;
 import javax.swing.ImageIcon;
