+++ lib/src/main/java/com/zetcode/model/bold.png
