+++ lib/src/main/java/com/zetcode/model/italic.png
