+++ lib/src/main/java/com/zetcode/model/strike.png
